==> P5_land_bidding_xgboost.ipynb <==
import pandas as pd
import numpy as np
from xgboost import XGBRegressor, DMatrix, cv, train, plot_tree
from sklearn.model_selection import train_test_split
from sklearn.metrics import mean_squared_error as mse, mean_absolute_percentage_error as mape, r2_score
import matplotlib.pylab as plt
import graphviz

# Hyperparameter tuning
from sklearn.model_selection import RepeatedKFold, cross_val_score, GridSearchCV, RandomizedSearchCV
from hyperopt import tpe, STATUS_OK, Trials, hp, fmin, STATUS_OK, space_eval

### Read in data

gls = pd.read_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1207_new_feat.csv')
gls = gls.sort_values(by=['year_launch', 'month_launch', 'date_launch'])

gls.columns

# select features
categories = ['region', 'zone', 'devt_class', 'source']
numeric = ['site_area_sqm', 'lease_term', 'gpr',
           'num_bidders', 'timediff_launch_to_close',
           'avg_dist_cbd', 'avg_dist_mrt', 'avg_num_bus',
           'avg_num_good_sch', 'proj_num_of_units', 'proj_max_floor',
           'num_of_nearby_completed_proj_200m', 'num_of_schools', 'year_launch']
cols = categories + numeric

### Pretreatment

x = pd.get_dummies(gls[cols])
y = gls.price_psm_real

dmat = DMatrix(data=x, label=y)

### Split train-test

x_train, x_test, y_train, y_test = train_test_split(x, y, shuffle=False, test_size=0.2)
train_data = DMatrix(x_train, label=y_train)
test_data = DMatrix(x_test, label=y_test)

### Set model and regress

# using train
params = {
    'max_depth': 6,
    'eta': 0.02,
    'objective': 'reg:squarederror',
}
xgb = XGBRegressor(objective ='reg:squarederror').fit(x_train, y_train)
xgb_reg = train(params=params, dtrain=train_data, num_boost_round=100)
pred_train = xgb_reg.predict(train_data)
mape_train = mape(y_train, pred_train)
pred_test = xgb_reg.predict(test_data)
mape_test = mape(y_test, pred_test)
print("MAPE train: %f" %mape_train, "MAPE test: %f" %mape_test, "MAPE test-train: %f" %(mape_test-mape_train), sep='\n')

r2_test, r2_train = r2_score(y_test, pred_test), r2_score(y_train, pred_train)
print(r2_test, r2_train, sep='\n')

test_check = pd.DataFrame({'y_actual': y_test, 'y_predict': pred_test})
train_check = pd.DataFrame({'actual': y_train, 'predict': pred_train})

xgb.get_params()

### Param tuning and random search

# Round 1
# define search space
param_space = {'max_depth': [4, 5, 6],
               'learning_rate': [0.01, 0.02, 0.05],
               'gamma': [0, 0.25, 1.0],
               'reg_lambda': [0, 1.0, 5.0]
               }
scoring = ['neg_mean_absolute_percentage_error']
kfold = RepeatedKFold(n_splits=10, n_repeats=3, random_state=42)
random_search = RandomizedSearchCV(estimator=xgb,
                                    param_distributions=param_space,
                                    n_iter=50,
                                    scoring=scoring,
                                    refit=scoring[0],
                                    n_jobs=-1,
                                    cv=kfold,
                                    verbose=1)

random_search_res = random_search.fit(x_train, y_train)
# Print the best score and the corresponding hyperparameters
print(f'The best score is {random_search_res.best_score_:.4f}')
# print('The best score standard deviation is', round(randome_search_res.cv_results_['std_test_recall'][randome_search_res.best_index_], 4))
print(f'The best hyperparameters are {random_search_res.best_params_}')

# Round 2
# define search space
param_space = {'max_depth': [4, 5, 6],
               'learning_rate': [0.01, 0.02, 0.05],
               'gamma': [0, 0.25, 1.0],
               'reg_lambda': [0, 1.0, 5.0]
               }
scoring = ['neg_mean_absolute_percentage_error']
kfold = RepeatedKFold(n_splits=10, n_repeats=3, random_state=42)
random_search = RandomizedSearchCV(estimator=xgb,
                                   param_distributions=param_space,
                                   n_iter=50,
                                   scoring=scoring,
                                   refit=scoring[0],
                                   n_jobs=-1,
                                   cv=kfold,
                                   verbose=1)

random_search_res = random_search.fit(x_train, y_train)
# Print the best score and the corresponding hyperparameters
print(f'The best score is {random_search_res.best_score_:.4f}')
# print('The best score standard deviation is', round(randome_search_res.cv_results_['std_test_recall'][randome_search_res.best_index_], 4))
print(f'The best hyperparameters are {random_search_res.best_params_}')

validation_res = cv(dtrain=train_data, params=params, nfold=3, num_boost_round=100,
                    early_stopping_rounds=10, metrics='rmse', as_pandas=True, seed=42)

validation_res['test-rmse-mean'].tail(1)

### Visualize

# xg_reg = train(params=params, dtrain=dmat, num_boost_round=100)
# plot_tree(xg_reg, num_trees=0)
# plt.rcParams['figure.figsize'] = [50, 10]
# plt.show()
