==> P24_joining_lat_lon.ipynb <==
import pandas as pd
import numpy as np
import hashlib

loc = pd.read_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_geo.csv')
loc['land_parcel_std'] = loc.land_parcel_std.apply(lambda x: x.replace('/', ' / '))

loc['parcel_id_text'] = loc.land_parcel_std + loc.latitude.astype(str) + loc.longitude.astype(str)
loc['land_parcel_id'] = loc.parcel_id_text.apply(lambda x: hashlib.sha256(x.encode('utf-8')).hexdigest())

loc['proj_name_raw'] = loc.proj_name_raw.apply(lambda x: x.lower().replace(' ', '-'))

y_cols = list(loc.columns)
y_cols.remove('parcel_id_text')
y_cols

### Join to main table

gls = pd.read_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1208.csv')

header = list(gls.columns)
header

selected_cols = [col for col in header if col not in y_cols]
selected_cols.append('sg_gls_id')
selected_cols

gls_selected = gls[selected_cols]

header = header[:13] + ['latitude', 'longitude'] + header[13:]

gls_w_loc = pd.merge(loc[y_cols], gls_selected, on='sg_gls_id', how='left')[header]

gls_w_loc.to_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1209.csv', index = False)

gls_edit = pd.read_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1212.csv')

gls_edit['gls_text_id'] = gls_edit.latitude.astype(str) + gls_edit.longitude.astype(str) + gls_edit.award_month_index.astype(str)
gls_edit.insert(loc=0, column='sg_gls_id_new', value=gls_edit.gls_text_id.apply(lambda x: hashlib.sha256(x.encode('utf-8')).hexdigest()))

gls_edit.to_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1212_v2.csv', index = False)

gls_v3 = pd.read_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1212_v3.csv')
gls_v3['gls_text_id'] = gls_v3.land_parcel_std + gls_v3.latitude.astype(str) + gls_v3.longitude.astype(str) + gls_v3.award_month_index.astype(str) + gls_v3.site_area_sqm.astype(str)
gls_v3.insert(loc=0, column='sg_gls_id_new', value=gls_v3.gls_text_id.apply(lambda x: hashlib.sha256(x.encode('utf-8')).hexdigest()))

gls_v3.to_csv(r'G:\REA\Working files\land-bidding\land_sales_full_data\feature eng\gls_1212_v4.csv', index = False)
